Create base class for Nek mesh generators. Refs #501
